feat(thumbnailProvider): remove default LocalThumbnailProvider
and require a dedicated thumbnail provider to be configured

in order to make dependency to sharp optional.
(sharp is quite big and in FaaS envs we want to use s.th. else)

BREAKING CHANGE:
`customThumbnailProvider` option is renamed to `thumbnailProvider` and not longer optional

Migration:
install `@cotype/local-thumbnail-provider` and use it as thumbnailProvider

fix https://github.com/cotype/core/issues/1